+++ benchmarks/stringify-stream-conformance.js
@@ -1,4 +1,5 @@
 import { inspect } from 'node:util';
+import { Readable } from 'node:stream';
 import chalk from 'chalk';
 import bfj from 'bfj';
 import JsonStreamStringify from 'json-stream-stringify';
@@ -43,6 +44,7 @@ async function test(createStream, value, replacer, space) {
 
 const streamFactories = {
     [selfPackageJson.name]: (value, replacer, space) => jsonExt.stringifyStream(value, replacer, space),
+    [selfPackageJson.name + '/strigifyChunked']: (value, replacer, space) => Readable.from(jsonExt.stringifyChunked(value, replacer, space)),
     'bfj': (value, replacer, space) => bfj.streamify(value, { space }),
     'json-stream-stringify': (value, replacer, space) => new JsonStreamStringify(value, replacer, space)
 };