+++ src/index.js
@@ -1,4 +1,5 @@
 export { version } from './version.js';
 export { stringifyInfo } from './stringify-info.js';
 export { stringifyStream } from './stringify-stream.js';
+export { stringifyChunked } from './stringify-chunked.js';
 export { parseChunked } from './parse-chunked.js';